Replace regex with hand-crafted parser
Improve performance by replacing the generic regular expression with a
hand-crafted parser. Parsing is fast, but dumb and fragile, and will
probably break when the input format changes. For instance, it assumes
that the only occurrences of { and } are the start and end of JSON data,
which happens to be true at the time of writing.

According to hyperfine, this commit improves runtime (using the full
module-info.json file) from 473.1 ms to 87.8 ms.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>